+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+
+* Added GestureDetector. Soft keyboard on device is dismissible by
+touching the screen
+
 ## 1.0.4
 
 Library is working like a charm in this version